feat(view-stops-in-map): migrate processing logic to util file
